requirements.txt
certifi==2019.11.28
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.3.0
numpy==1.16.5
olefile==0.46
pandas==0.25.1
Pillow==6.2.1
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
wincertstore==0.2
